refactor(api): relocate suggest_prompts to client.notebooks (pre-release)
GeneratePromptSuggestions (otmP3b) is a general notebook-prompt endpoint, not
chat-specific: the `mode` arg selects the product surface (chat / critique /
audio-debate / quiz / flashcards — live-characterized, stable per mode). Move
the public method client.chat.suggest_prompts -> client.notebooks.suggest_prompts
before release, so the "chat" misnomer never ships. Not a breaking change — the
method was never in a release (api-compat audit clean).

- method: ChatAPI -> NotebooksAPI (calls self.get_source_ids directly)
- builder/constants: _chat/payloads.py -> _notebook_payloads.py
- row adapter: prompt-suggestion bits -> new _row_adapters/notebooks.py
- type: _types/chat.PromptSuggestion -> _types/notebooks (public names unchanged)
- tests / cassette / integration / docs renamed; mode stays an int (1..9)
- RPC plumbing (RPCMethod.SUGGEST_PROMPTS, golden, cassette body, idempotency)
  unchanged — it is RPC-keyed and namespace-agnostic
- docstrings corrected: modes are stable surface selectors (live A/B proof),
  NOT interchangeable; default 4 = chat questions
