+++ docs/rpc-reference.md
@@ -47,7 +47,7 @@
 | `hPTbtc` | GET_LAST_CONVERSATION_ID | Get most recent conversation ID | `_chat/api.py` |
 | `khqZz` | GET_CONVERSATION_TURNS | Get Q&A turns for a conversation | `_chat/api.py` |
 | `J7Gthc` | DELETE_CONVERSATION | Delete a conversation (web UI's "Delete history") | `_chat/api.py` |
-| `otmP3b` | SUGGEST_PROMPTS | Get AI-suggested questions/prompts to ask a notebook | `_chat/api.py` |
+| `otmP3b` | SUGGEST_PROMPTS | Get AI-suggested prompts for a notebook | `_notebooks.py` |
 | `CYK0Xb` | CREATE_NOTE | Create a note (placeholder) | `_notes.py` |
 | `cYAfTb` | UPDATE_NOTE | Update note content/title | `_notes.py` |
 | `AH0mwd` | DELETE_NOTE | Delete a note | `_notes.py` |
@@ -953,13 +953,14 @@ is signaled by the absence of an error — there is no return payload.
 
 ### RPC: SUGGEST_PROMPTS (otmP3b)
 
-**Source:** `_chat/api.py::suggest_prompts()`
+**Source:** `_notebooks.py::NotebooksAPI.suggest_prompts()`
 
-Returns AI-suggested questions/prompts to ask a notebook (the live
-`GeneratePromptSuggestions` method). Each suggestion pairs a short title with a
-ready-to-send multi-line instruction string. Shape live-verified on the
-consumer/labs cohort (issue #1612) — the backend serves it regardless of the
-web UI's experiment flag.
+Returns AI-suggested prompts for a notebook (the live
+`GeneratePromptSuggestions` method) — a general notebook-prompt endpoint whose
+`mode` selects the product surface (default `4` suggests chat questions). Each
+suggestion pairs a short title with a ready-to-send multi-line instruction
+string. Shape live-verified on the consumer/labs cohort (issue #1612) — the
+backend serves it regardless of the web UI's experiment flag.
 
 ```python
 params = [
