+++ src/notebooklm/_chat/__init__.py
@@ -7,13 +7,13 @@ collision; its public names are re-exported here so existing references such as
 ``from notebooklm._chat import ChatAPI`` keep resolving unchanged.
 
 No package-init import cycle exists: the dependency direction is strictly one-way
-(``api`` imports ``notes``/``payloads``/``wire``/``transport``; none of the helpers
+(``api`` imports ``notes``/``wire``/``transport``; none of the helpers
 import ``api`` or this package ``__init__``). Importing ``api`` therefore pulls in the
 helpers it needs regardless of the order of the aggregating ``from . import ...`` line
 below, which ruff's import sorter keeps alphabetised.
 """
 
-from . import api, notes, payloads, transport, wire
+from . import api, notes, transport, wire
 from .api import ChatAPI, _extract_next_turn_content
 
 # ``_extract_next_turn_content`` is private-by-name but intentionally re-exported
@@ -24,7 +24,6 @@ from .api import ChatAPI, _extract_next_turn_content
 __all__ = [
     "api",
     "notes",
-    "payloads",
     "transport",
     "wire",
     "ChatAPI",
