+++ src/notebooklm/_notebooks.py
@@ -10,6 +10,11 @@ from ._notebook_metadata import (
     NotebookSourceLister,
     create_default_source_lister,
 )
+from ._notebook_payloads import (
+    _PROMPT_SUGGESTIONS_DEFAULT_MODE,
+    build_prompt_suggestions_params,
+)
+from ._row_adapters.notebooks import PromptSuggestionRow, unwrap_prompt_suggestions
 from ._row_adapters.sources import SourceRow
 from ._runtime.contracts import RpcCaller
 from ._settings import build_get_user_settings_params, extract_account_limits
@@ -24,9 +29,17 @@ from .exceptions import (
     RateLimitError,
     RPCError,
     ServerError,
+    ValidationError,
 )
 from .rpc import RPCMethod, safe_index
-from .types import AccountLimits, Notebook, NotebookDescription, NotebookMetadata, SuggestedTopic
+from .types import (
+    AccountLimits,
+    Notebook,
+    NotebookDescription,
+    NotebookMetadata,
+    PromptSuggestion,
+    SuggestedTopic,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -349,6 +362,87 @@ class NotebooksAPI:
             )
 
         return source_ids
+
+    async def suggest_prompts(
+        self,
+        notebook_id: str,
+        *,
+        source_ids: list[str] | None = None,
+        mode: int = _PROMPT_SUGGESTIONS_DEFAULT_MODE,
+        query: str | None = None,
+    ) -> list[PromptSuggestion]:
+        """Get AI-suggested prompts for a notebook.
+
+        Backed by ``GeneratePromptSuggestions`` (``otmP3b``): a *general*
+        notebook-prompt endpoint whose ``mode`` selects the product surface to
+        suggest for. With the default ``mode=4`` the server suggests chat
+        questions to ask :meth:`ChatAPI.ask`; other modes target other surfaces
+        (critique, audio/debate, quiz, flashcards). The server returns a short
+        list of ``{title, prompt}`` suggestions, each ``prompt`` a ready-to-send
+        multi-line instruction.
+
+        Args:
+            notebook_id: The notebook to suggest prompts for.
+            source_ids: Source ids to scope the suggestions to. ``None``
+                (default) uses **all** of the notebook's sources.
+            mode: The required ``C0`` int "mode/surface" enum, inclusive range
+                ``1..9`` (``0`` / omitted makes the server return ``INTERNAL``).
+                The suggestions are LLM-generated (non-deterministic) but their
+                *framing* is a stable function of ``mode``: it selects the product
+                surface the prompts are written for. Default ``4`` = general "ask
+                about the content" questions (the web chat surface's own default);
+                ``5`` = critique/evaluate; ``6`` = audio/debate; ``8`` = quiz;
+                ``9`` = flashcards; ``1-3`` and ``7`` track ``4``. Stays a plain
+                int, not a named enum, since the bundle exposes the values but not
+                Google's member names. See ``_PROMPT_SUGGESTIONS_DEFAULT_MODE`` for
+                the full bundle + live-probe table.
+            query: Optional free-text steer for the kind of prompts to suggest.
+                An empty / whitespace-only string is treated as no steer.
+
+        Returns:
+            A list of :class:`~notebooklm.types.PromptSuggestion`. An empty /
+            degenerate server response yields ``[]`` (suggestions are
+            best-effort UI sugar — an absent payload does not raise).
+
+        Raises:
+            ValidationError: if ``mode`` is outside the inclusive ``1..9`` range
+                (caught before any network call, so a bad mode never costs an
+                RPC).
+
+        .. versionadded:: 0.8.0
+        """
+        logger.debug("Suggesting prompts for notebook %s (mode=%d)", notebook_id, mode)
+        # Validate the mode up front (before the source-id fetch) so a bad value
+        # fails fast without a wasted round-trip; the builder's ValueError is
+        # re-raised as the public ValidationError for a uniform error contract.
+        try:
+            build_prompt_suggestions_params(notebook_id, [], mode=mode)
+        except ValueError as exc:
+            raise ValidationError(str(exc)) from exc
+        if source_ids is None:
+            source_ids = await self.get_source_ids(notebook_id)
+
+        params = build_prompt_suggestions_params(notebook_id, source_ids, mode=mode, query=query)
+        result = await self._rpc.rpc_call(
+            RPCMethod.SUGGEST_PROMPTS,
+            params,
+            source_path=f"/notebook/{notebook_id}",
+            allow_null=True,
+        )
+
+        rows = unwrap_prompt_suggestions(result, source="suggest_prompts")
+        # ``is_well_formed`` only gates on row LENGTH (>= 2 slots), not on the
+        # field values, mirroring ``ReportSuggestionRow``: a length-ok row whose
+        # title/prompt degrade to "" (a non-string leaf) still maps to a
+        # ``PromptSuggestion("", "")``. Real traffic always carries string
+        # leaves, so this is a best-effort tolerance for a degenerate server
+        # payload, not an expected output — callers should not treat an empty
+        # title/prompt as meaningful.
+        return [
+            PromptSuggestion(title=row.title, prompt=row.prompt)
+            for row in map(PromptSuggestionRow, rows)
+            if row.is_well_formed
+        ]
 
     async def list(self) -> list[Notebook]:
         """List notebooks (most-recently-viewed first).
