+++ src/notebooklm/_types/chat.py
@@ -137,21 +137,3 @@ class AskResult:
     is_follow_up: bool
     references: list[ChatReference] = field(default_factory=list)
     raw_response: str = ""
-
-
-@dataclass(frozen=True)
-class PromptSuggestion:
-    """An AI-suggested question/prompt to ask a notebook.
-
-    Returned by :meth:`ChatAPI.suggest_prompts` (the ``otmP3b`` /
-    ``GeneratePromptSuggestions`` RPC). Each suggestion pairs a short,
-    human-readable ``title`` with a ready-to-send multi-line ``prompt`` that can
-    be passed straight to :meth:`ChatAPI.ask`.
-
-    Attributes:
-        title: Short label for the suggestion (e.g. ``"Professional Briefing"``).
-        prompt: The full multi-line instruction string to ask the notebook.
-    """
-
-    title: str
-    prompt: str
