+++ tests/_guardrails/test_golden_decode_coverage.py
@@ -89,7 +89,7 @@ _GOLDEN_VCR = "tests/integration/test_golden_decoded_vcr.py"
 _GOLDEN_EXPANSION = "tests/integration/test_golden_decoded_vcr_expansion.py"
 _COMPREHENSIVE = "tests/integration/test_vcr_comprehensive.py"
 _GAP_BACKFILL = "tests/integration/test_rpc_gap_backfill_vcr.py"
-_SUGGEST_PROMPTS_VCR = "tests/integration/test_chat_suggest_prompts_vcr.py"
+_SUGGEST_PROMPTS_VCR = "tests/integration/test_notebooks_suggest_prompts_vcr.py"
 
 GoldenPointer = tuple[str, str]
