+++ tests/cassettes/notebooks_suggest_prompts.yaml
