+++ tests/unit/test_notebooks_suggest_prompts.py
@@ -1,10 +1,10 @@
-"""Unit tests for ``ChatAPI.suggest_prompts`` (``otmP3b`` /
+"""Unit tests for ``NotebooksAPI.suggest_prompts`` (``otmP3b`` /
 ``GeneratePromptSuggestions``).
 
 Pins the wire contract (param-builder shape + ``source_path``) and the
 response decode, including the empty / degenerate payloads. No network: the
-``rpc`` collaborator and the source-id provider are narrow mocks, matching the
-``suggest_reports`` unwrap tests and the other ``ChatAPI`` unit tests.
+``rpc`` collaborator is a narrow mock and ``get_source_ids`` is patched, matching
+the ``suggest_reports`` unwrap tests and the other ``NotebooksAPI`` unit tests.
 """
 
 from __future__ import annotations
@@ -13,10 +13,8 @@ from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 
-from notebooklm._chat import ChatAPI
-from notebooklm._chat.payloads import build_prompt_suggestions_params
-from notebooklm._notebook_metadata import NotebookSourceIdProvider
-from notebooklm._runtime.contracts import LoopGuard, RpcCaller
+from notebooklm._notebooks import NotebooksAPI, build_prompt_suggestions_params
+from notebooklm._runtime.contracts import RpcCaller
 from notebooklm.exceptions import ValidationError
 from notebooklm.rpc import RPCMethod
 from notebooklm.types import PromptSuggestion
@@ -38,23 +36,15 @@ def mock_rpc() -> MagicMock:
 
 
 @pytest.fixture
-def mock_notebooks() -> MagicMock:
-    """Source-id provider returning two ids for the default-all-sources path."""
-    return MagicMock(
-        spec=NotebookSourceIdProvider,
-        get_source_ids=AsyncMock(return_value=["src_a", "src_b"]),
-    )
+def api(mock_rpc: MagicMock) -> NotebooksAPI:
+    """A ``NotebooksAPI`` whose ``get_source_ids`` returns two ids by default.
 
-
-@pytest.fixture
-def api(mock_rpc: MagicMock, mock_notebooks: MagicMock) -> ChatAPI:
-    return ChatAPI(
-        rpc=mock_rpc,
-        transport=MagicMock(),
-        reqid=MagicMock(),
-        loop_guard=MagicMock(spec=LoopGuard),
-        notebooks=mock_notebooks,
-    )
+    The default-all-sources path resolves ids through ``get_source_ids``; tests
+    that pass explicit ``source_ids`` assert it is never awaited.
+    """
+    notebooks = NotebooksAPI(mock_rpc)
+    notebooks.get_source_ids = AsyncMock(return_value=["src_a", "src_b"])  # type: ignore[method-assign]
+    return notebooks
 
 
 # ---------------------------------------------------------------------------
@@ -117,7 +107,7 @@ class TestBuildPromptSuggestionsParams:
 class TestSuggestPrompts:
     @pytest.mark.asyncio
     async def test_decodes_rows_to_prompt_suggestions(
-        self, api: ChatAPI, mock_rpc: MagicMock
+        self, api: NotebooksAPI, mock_rpc: MagicMock
     ) -> None:
         result = await api.suggest_prompts("nb_xyz", source_ids=["s1"], mode=4)
         assert result == [
@@ -132,7 +122,7 @@ class TestSuggestPrompts:
         ]
 
     @pytest.mark.asyncio
-    async def test_sends_expected_rpc(self, api: ChatAPI, mock_rpc: MagicMock) -> None:
+    async def test_sends_expected_rpc(self, api: NotebooksAPI, mock_rpc: MagicMock) -> None:
         await api.suggest_prompts("nb_xyz", source_ids=["s1", "s2"], mode=7, query="q")
 
         mock_rpc.rpc_call.assert_awaited_once()
@@ -144,35 +134,33 @@ class TestSuggestPrompts:
 
     @pytest.mark.asyncio
     async def test_defaults_to_all_sources_when_source_ids_none(
-        self, api: ChatAPI, mock_rpc: MagicMock, mock_notebooks: MagicMock
+        self, api: NotebooksAPI, mock_rpc: MagicMock
     ) -> None:
         await api.suggest_prompts("nb_xyz")
 
-        mock_notebooks.get_source_ids.assert_awaited_once_with("nb_xyz")
+        api.get_source_ids.assert_awaited_once_with("nb_xyz")  # type: ignore[attr-defined]
         # The resolved source ids are wrapped into f3 of the params.
         _, _, source_block, *_ = mock_rpc.rpc_call.call_args.args[1]
         assert source_block == [["src_a"], ["src_b"]]
 
     @pytest.mark.asyncio
-    async def test_explicit_source_ids_skip_resolution(
-        self, api: ChatAPI, mock_notebooks: MagicMock
-    ) -> None:
+    async def test_explicit_source_ids_skip_resolution(self, api: NotebooksAPI) -> None:
         await api.suggest_prompts("nb_xyz", source_ids=["only"])
-        mock_notebooks.get_source_ids.assert_not_awaited()
+        api.get_source_ids.assert_not_awaited()  # type: ignore[attr-defined]
 
     @pytest.mark.asyncio
     @pytest.mark.parametrize("mode", [0, 10])
     async def test_out_of_range_mode_raises_before_any_network(
-        self, api: ChatAPI, mock_rpc: MagicMock, mock_notebooks: MagicMock, mode: int
+        self, api: NotebooksAPI, mock_rpc: MagicMock, mode: int
     ) -> None:
         """A bad mode raises ValidationError before the source fetch or the RPC."""
         with pytest.raises(ValidationError, match="1..9"):
             await api.suggest_prompts("nb_xyz", mode=mode)
-        mock_notebooks.get_source_ids.assert_not_awaited()
+        api.get_source_ids.assert_not_awaited()  # type: ignore[attr-defined]
         mock_rpc.rpc_call.assert_not_awaited()
 
     @pytest.mark.asyncio
-    async def test_single_suggestion_wrapped(self, api: ChatAPI, mock_rpc: MagicMock) -> None:
+    async def test_single_suggestion_wrapped(self, api: NotebooksAPI, mock_rpc: MagicMock) -> None:
         """A wrapped single-row envelope (``[[[t, p]]]``) decodes one suggestion."""
         mock_rpc.rpc_call.return_value = [[["Solo", "\n- One prompt."]]]
         result = await api.suggest_prompts("nb_xyz", source_ids=["s1"])
@@ -181,13 +169,13 @@ class TestSuggestPrompts:
     @pytest.mark.asyncio
     @pytest.mark.parametrize("payload", [None, [], [[]], "unexpected", [None]])
     async def test_degenerate_payloads_yield_empty(
-        self, api: ChatAPI, mock_rpc: MagicMock, payload: object
+        self, api: NotebooksAPI, mock_rpc: MagicMock, payload: object
     ) -> None:
         mock_rpc.rpc_call.return_value = payload
         assert await api.suggest_prompts("nb_xyz", source_ids=["s1"]) == []
 
     @pytest.mark.asyncio
-    async def test_malformed_rows_are_skipped(self, api: ChatAPI, mock_rpc: MagicMock) -> None:
+    async def test_malformed_rows_are_skipped(self, api: NotebooksAPI, mock_rpc: MagicMock) -> None:
         """Short / non-list rows degrade out; well-formed rows survive."""
         mock_rpc.rpc_call.return_value = [
             [
